c: directory "shared" is no longer used
