+++ client/src/App.tsx
@@ -1,7 +1,7 @@
 import React, { lazy, Suspense } from 'react'
 import AppLayout from '@/components/AppLayout'
 import Fallback from '@/components/Fallback'
-import PrivateRoute from '@/shared/PrivateRoute'
+import PrivateRoute from '@/components/PrivateRoute'
 import { Routes, Route, Navigate } from 'react-router-dom'
 import { useWindowCSSRatio } from '@/hooks'
