+++ client/src/components/ImagePlaceholder/styles.css
