+++ client/src/components/MiniPlayer/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
-import SvgIcon from '@/shared/SvgIcon'
-import IconButton from '@/shared/IconButton'
+import SvgIcon from '@/components/ui/SvgIcon'
+import IconButton from '@/components/ui/IconButton'
 import PlayIcon from '@/assets/svg/play.svg?react'
 import PauseIcon from '@/assets/svg/pause.svg?react'
-import ImagePlaceholder from '@/shared/ImagePlaceholder'
+import ImagePlaceholder from '@/components/ImagePlaceholder'
 import { useReactiveVar } from '@apollo/client'
 import { currentTrackVar } from '@/apollo/cache/variables'
