+++ client/src/components/Player/Controls/index.tsx
@@ -4,7 +4,7 @@ import PauseIcon from '@/assets/svg/pause.svg?react'
 import ArrowAltIcon from '@/assets/svg/arrow-alt.svg?react'
 import RepeatIcon from '@/assets/svg/repeat.svg?react'
 import RandomIcon from '@/assets/svg/random.svg?react'
-import SvgIcon from '@/shared/SvgIcon'
+import SvgIcon from '@/components/ui/SvgIcon'
 import {
   Container,
   PlayButtonBox,
