+++ client/src/components/Player/TopControls/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SvgIcon from '@/shared/SvgIcon'
+import SvgIcon from '@/components/ui/SvgIcon'
 import ArrowIcon from '@/assets/svg/angle-arrow.svg?react'
 import MusicListIcon from '@/assets/svg/list-music.svg?react'
 import { useTranslation } from 'react-i18next'
