+++ client/src/components/PlaylistItem/index.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef } from 'react'
 import styled from 'styled-components'
-import ScaleLoader from '@/shared/Loader'
-import ImagePlaceholder from '@/shared/ImagePlaceholder'
+import ScaleLoader from '@/components/ui/Loader'
+import ImagePlaceholder from '@/components/ImagePlaceholder'
 import { timeFormater } from '@/utils/helpers'
 import { PlaylistMutations, SoundtrackMutations } from '@/apollo/cache/mutations'
 import type { Playlist } from '@/apollo/cache/types'
