+++ client/src/components/ui/Button/index.tsx
@@ -1,6 +1,6 @@
-import React from 'react'
+import React, { forwardRef } from 'react'
 import styled, { css } from 'styled-components'
-import Loader from '@/shared/Loader'
+import Loader from '@/components/ui/Loader'
 
 enum ButtonColor {
   primary = 'primary',
@@ -11,27 +11,42 @@ enum ButtonColor {
 
 interface BottonProps {
   to?: string
-  color?: 'primary' | 'secondary' | 'success' | 'danger'
   type?: 'button' | 'reset' | 'submit'
+  color?: 'primary' | 'secondary' | 'success' | 'danger'
   children: React.ReactNode
+  startIcon?: React.ReactNode
+  endIcon?: React.ReactNode
   loading?: boolean
   disabled?: boolean
-  withoutShadow?: boolean
+  fullWidth?: boolean
+  withShadow?: boolean
+  tabIndex?: number
+  ref?: React.RefObject<HTMLButtonElement> | null
   onClick?(event: React.MouseEvent<HTMLButtonElement>): void
 }
 
 interface BaseButtonProps {
-  $withoutShadow: boolean
+  $fullWidth?: boolean
+  $withShadow: boolean
   $color: ButtonColor
 }
+
+const LoaderWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+`
 
 const BaseButton = styled.button<BaseButtonProps>`
   display: inline-flex;
   align-items: center;
   justify-content: center;
-  padding: 20px;
-  border-radius: 50%;
-  border: none;
+  width: ${({ $fullWidth }) => ($fullWidth ? '100%' : 'auto')};
+  font-size: 18px;
+  line-height: 18px;
+  font-weight: 500;
+  padding: 15px 30px;
+  border-radius: 8px;
+  border: 1px solid;
   transition: all 0.3s;
   -webkit-tap-highlight-color: transparent;
   &:disabled {
@@ -39,20 +54,21 @@ const BaseButton = styled.button<BaseButtonProps>`
     pointer-events: initial;
     opacity: 0.7;
   }
-  ${({ $withoutShadow }) =>
-    $withoutShadow
-      ? css`
-          box-shadow: 0px 8px 17px rgba(0, 0, 0, 0.3);
-        `
-      : css`
-          box-shadow: none;
-        `}
+  &:focus {
+    outline: none;
+  }
+  ${({ $withShadow }) =>
+    $withShadow &&
+    css`
+      box-shadow: 0px 8px 17px rgba(0, 0, 0, 0.3);
+    `}
   ${({ $color, theme }) => {
     switch ($color) {
       case ButtonColor.primary:
         return css`
           color: ${theme.colors.primary};
           background-color: #363636;
+          border-color: #6a6a6a;
           &:hover:not(:disabled) {
             cursor: pointer;
             background-color: #323232;
@@ -89,24 +105,40 @@ const BaseButton = styled.button<BaseButtonProps>`
   }}
 `
 
-const IconButton: React.FC<BottonProps> = ({
-  loading,
-  children,
-  color = ButtonColor.primary,
-  withoutShadow = false,
-  type = 'button',
-  ...otherProps
-}: BottonProps) => {
-  return (
-    <BaseButton
-      $color={color as ButtonColor}
-      $withoutShadow={!withoutShadow}
-      type={type}
-      {...otherProps}
-    >
-      {loading ? <Loader dark={color !== ButtonColor.primary} /> : children}
-    </BaseButton>
-  )
-}
+const ButtonText = styled.span``
 
-export default IconButton
+const Button: React.ForwardRefRenderFunction<HTMLButtonElement, BottonProps> = (
+  {
+    loading,
+    children,
+    color = ButtonColor.primary,
+    withShadow = false,
+    startIcon,
+    endIcon,
+    type = 'button',
+    fullWidth,
+    ...otherProps
+  },
+  ref
+) => (
+  <BaseButton
+    ref={ref}
+    $color={color as ButtonColor}
+    $withShadow={withShadow}
+    $fullWidth={fullWidth}
+    type={type}
+    {...otherProps}
+  >
+    {!loading && startIcon}
+    {loading ? (
+      <LoaderWrapper>
+        <Loader dark={color !== ButtonColor.primary} />
+      </LoaderWrapper>
+    ) : (
+      <ButtonText>{children}</ButtonText>
+    )}
+    {!loading && endIcon}
+  </BaseButton>
+)
+
+export default forwardRef(Button)
