+++ client/src/components/ui/SvgIcon/index.tsx
