+++ client/src/pages/Home/NowPlaying/index.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from 'react'
-import SvgIcon from '@/shared/SvgIcon'
+import SvgIcon from '@/components/ui/SvgIcon'
 import PlusIcon from '@/assets/svg/plus.svg?react'
 import TrashIcon from '@/assets/svg/trash.svg?react'
 import ShareIcon from '@/assets/svg/share.svg?react'
 import CopyIcon from '@/assets/svg/copy.svg?react'
 import CheckIcon from '@/assets/svg/check-circle.svg?react'
-import ImagePlaceholder from '@/shared/ImagePlaceholder'
+import ImagePlaceholder from '@/components/ImagePlaceholder'
 import { useReactiveVar } from '@apollo/client'
 import { currentTrackVar, userVar } from '@/apollo/cache/variables'
 import { useTranslation } from 'react-i18next'
