+++ client/src/pages/Upload/Audio/index.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, useRef } from 'react'
 import Dropzone from './Dropzone'
-import SvgIcon from '@/shared/SvgIcon'
+import SvgIcon from '@/components/ui/SvgIcon'
 import StepControls from '../StepControls'
 import PlayIcon from '@/assets/svg/play.svg?react'
 import PauseIcon from '@/assets/svg/pause.svg?react'
-import Checkbox from '@/shared/FormFields/Checkbox'
+import Checkbox from '@/components/ui/Checkbox'
 import AudioPlayer from '@/player'
 import { timeFormater } from '@/utils/helpers'
 import { useFormikContext } from 'formik'
