+++ client/src/pages/Upload/Info/index.tsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react'
 import styled from 'styled-components'
-import Button from '@/shared/Button'
-import TextInput from '@/shared/FormFields/TextInput'
+import Button from '@/components/ui/Button'
+import TextInput from '@/components/ui/TextInput'
 import blushEmoji from '@/assets/rastr/blush.png'
 import { useFormikContext } from 'formik'
 import { useTranslation } from 'react-i18next'
