+++ client/src/pages/Upload/StepControls.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-import SvgIcon from '@/shared/SvgIcon'
-import Button from '@/shared/Button'
-import IconButton from '@/shared/IconButton'
+import SvgIcon from '@/components/ui/SvgIcon'
+import Button from '@/components/ui/Button'
+import IconButton from '@/components/ui/IconButton'
 import ArrowIcon from '@/assets/svg/angle-arrow.svg?react'
 import { useFormikContext } from 'formik'
 import { useTranslation } from 'react-i18next'
